+++ src_graph/GRAPH_GraphicalFunctions.h
@@ -176,7 +176,6 @@ template <typename Tgr> Tgr GRAPH_Read(std::istream &is) {
   is >> nbVert;
   Tgr eGR(nbVert);
   int SetHasColor_i;
-  ;
   is >> SetHasColor_i;
   bool SetHasColor = SetHasColor_i;
   eGR.SetHasColor(SetHasColor);
@@ -699,7 +698,7 @@ int L1_distance(std::vector<int> const &V1, std::vector<int> const &V2) {
 
 template <typename Tgr>
 std::vector<MyMatrix<int>> GRAPH_S_Embedding(Tgr const &GR, size_t const &s_sz,
-                                             long const &MaxIter, long &iter) {
+                                             size_t const &MaxIter, size_t &iter) {
   int s = static_cast<int>(s_sz);
   iter = 0;
   if (!IsSimpleGraph(GR)) {
